% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aprx_mi.R
\name{aprx_mi}
\alias{aprx_mi}
\title{Approximation of the mutual information between two variables}
\usage{
aprx_mi(data, n, x, y, logBase = 2, smth = 1e-10)
}
\arguments{
\item{data}{A categorical dataset.}

\item{n}{Sample size.}

\item{x}{A random variable X.}

\item{y}{A random variable Y.}

\item{logBase}{The base of logarithm. Default is 2.}

\item{smth}{A small value to get rid of 0 occurence.}
}
\description{
This function approximates the mutual information between two variables based
on the given data.
}
