% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_probs_adaptive.R
\name{cond_probs_adaptive}
\alias{cond_probs_adaptive}
\title{An auxiliary function to mml_fixed_str()}
\usage{
cond_probs_adaptive(data, arities, sampleSize, targetIndex, probsMtx,
  curIndex, curPaIndices)
}
\arguments{
\item{data}{A dataset whost variables are in numeric/integer format. Any categorical variables must be
converted into numeric/integer first.}

\item{arities}{A vector of variable arities in data, in the same order as the column names of data.}

\item{sampleSize}{The sample size. That is, the number of rows of data.}

\item{targetIndex}{The target node's index in vars, whose Markov blanket we are interested in.}

\item{probsMtx}{probsMtx}

\item{curIndex}{The current variable's index. The current variable can either be the target or
a different variable.}

\item{curPaIndices}{The indices of the current variable's parents.}
}
\value{
The function outputs a matrix of conditional probabilities with dimension arity(target) by
sample size.
}
\description{
This function calculates conditional probability at each data point for a variable if it is the
target or it has at least one parent.
}
