% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{conf_int}
\alias{conf_int}
\title{A function to calculate the confidence interval}
\usage{
conf_int(v, coefficient = 1.96)
}
\arguments{
\item{v}{A vector of statistics.}

\item{coefficient}{Default value is 1.96.}
}
\description{
This is a function to calculate the confidence interval of a given vector of statistics. The default coefficient 
1.96.
}
