% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_occurance.R
\name{count_occurance}
\alias{count_occurance}
\title{A function to count occurances of variable values}
\usage{
count_occurance(data, arities)
}
\arguments{
\item{data}{A categorical (nominal or ordinal) data set.}

\item{arities}{A list of arities for variables.}
}
\description{
This function store the occurance of each value of each node in a list in order to speed up the calculation
of mml_cpt.
}
