% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag2matrix.R
\name{dag2matrix}
\alias{dag2matrix}
\title{A function to convert a bnlearn format dag to its adjacency matrix}
\usage{
dag2matrix(dag)
}
\arguments{
\item{dag}{A dag or partial dag in bnlearn format.}
}
\description{
This function converts a bnlearn format dag to its adjacency matrix. The dag can be either directed or partially
directed.
}
