% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directing_tree.R
\name{directing_tree}
\alias{directing_tree}
\title{Directing a tree}
\usage{
directing_tree(tree, root)
}
\arguments{
\item{tree}{A tree that is saved in its adjacency matrix format.}

\item{root}{A user specified root.}
}
\description{
This function gives directions to a tree. It requires a root to be specified 
by users.
}
