% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dne2bnlearn.R
\name{dne2bnlearn}
\alias{dne2bnlearn}
\title{Convert netica's dne file to bnlearn dag format}
\usage{
dne2bnlearn(dneFile)
}
\arguments{
\item{dneFile}{a DAG saved in dne format.}
}
\description{
This is a function to convert the dag saved in netica's dne format (particularly
for CaMML's output) to bnlearn's dag format. The function has dependency
on readr::read_file to read dne file.
}
