% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_dist_dag.R
\name{edit_dist_dag}
\alias{edit_dist_dag}
\title{A function to calculate the edit distance from the learned to the true DAG}
\usage{
edit_dist_dag(dagL, dagT)
}
\arguments{
\item{dagL}{The learned DAG.}

\item{dagT}{The true DAG.}
}
\description{
The true and learned DAGs must be stored as adjacent matrix. The output is a
list consists of false positive, false negative, precision, recall and edit
distance. Edit distance between two DAGs is defined as the minimum number of
arc deletion and addition to turn dagL to dagT.
}
