% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_dist_mb.R
\name{edit_dist_mb}
\alias{edit_dist_mb}
\title{A function to calculate the edit distance from the learned Markov blanket to the true Markov blanket}
\usage{
edit_dist_mb(mbTrue, mbLearned, vars, target)
}
\arguments{
\item{mbTrue}{The true Markov blanket}

\item{mbLearned}{The learned Markov blanket}

\item{vars}{A vector of all variables.}

\item{target}{The target node.}
}
\description{
This is a function to calculate the edit distance from the learned Markov blanket to the true 
Markov blanket. Edit distance is defined as the number of operations required to get from one to the 
other. If both true and learned are empty, the edit distance is 0; if only one of them is 
empty, the edit distance is the size of the other; if none of them is empty, the edit distance is 
the false positives plus the false negatives.
}
