% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2numeric.R
\name{factor2numeric}
\alias{factor2numeric}
\title{Auxiliary function for mmlLogit}
\usage{
factor2numeric(data)
}
\arguments{
\item{data}{}
}
\description{
This function converts variable levels from strings to numeric starting from 0. For example, if a
varaible has levels A, B and C, these three levels will be converted to 0, 1 and 2 respectively.
}
