% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parents_instantiation_indices.R
\name{get_parents_instantiation_indices}
\alias{get_parents_instantiation_indices}
\title{Auxiliary function to mml_cpt()}
\usage{
get_parents_instantiation_indices(arities, numParents, parentsIndices,
  numParentsInstantiations, index)
}
\arguments{
\item{arities}{arities}

\item{numParents}{numParents}

\item{parentsIndices}{parentsIndices}

\item{numParentsInstantiations}{numParentsInstantiations}

\item{index}{index}
}
\description{
Auxiliary function to mml_cpt()
}
