% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob_sign.R
\name{get_prob_sign}
\alias{get_prob_sign}
\title{Auxiliary to fim_nb()}
\usage{
get_prob_sign(data)
}
\arguments{
\item{data}{A binary data set.}

\item{xIndices}{A vector of input variables indices.}
}
\description{
This function makes a matrix of 1 and -1 that is used to change sign of pik0 and pik1 when calculating 
FIM. When x[i, k] = 1 (i.e., 1st level), then 1, meaning no change; when x[i, k] = 2 (i.e., 2nd level), 
then -1, meaning change sign, since this parameter is expressed by 1 - p(xi = 1|y=1 or 0).
}
