% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_logit.R
\name{glm_logit}
\alias{glm_logit}
\title{A function to estimate logit model parameters}
\usage{
glm_logit(data, x, y)
}
\arguments{
\item{data}{}

\item{x}{A vector of input variables with any length. For an empty input variable, set x = c().}

\item{y}{The output/target variable.}
}
\description{
This function estimates binary logit model parameters using the build in function glm(). The output 
variable must be binary, whilst the input variable can be multinomial.
}
