% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_indices.R
\name{intersect_indices}
\alias{intersect_indices}
\title{Auxiliary function to mml_with_parents()}
\usage{
intersect_indices(numParents, parentsIndecies, indexListPerNodePerValue,
  potentialCombination)
}
\arguments{
\item{numParents}{numParents}

\item{parentsIndecies}{parentsIndecies}

\item{indexListPerNodePerValue}{indexListPerNodePerValue}

\item{potentialCombination}{potentialCombination}
}
\description{
This is an auxiliary function to mml_with_parents() to calculate the intersection indices.
}
