% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_tau_distance.R
\name{kendall_tau_distance}
\alias{kendall_tau_distance}
\title{Kendall tau distance}
\usage{
kendall_tau_distance(a, b)
}
\arguments{
\item{a}{A random vector or array.}

\item{b}{A random vector or array.
real value between 0 and 1, closed interval.}
}
\description{
Kendall tau distance is the number of pairs of nodes that are in a reversed
order in an array a and an array b.
}
