% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_extraction.R
\name{local_extraction}
\alias{local_extraction}
\title{A function that extract the local structure over a given set of variables from the true DAG}
\usage{
local_extraction(dag, varSubset)
}
\arguments{
\item{dag}{The true DAG.}

\item{varSubset}{A subset of variables whose local structure we are interested in.}
}
\description{
A function that extract the local structure over a given set of variables from the true DAG
}
