% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_factorial.R
\name{log_factorial}
\alias{log_factorial}
\title{A faster function to calculate log factorial of a non-negative integer}
\usage{
log_factorial(logFactorialSheet, n, base)
}
\arguments{
\item{logFactorialSheet}{A pre-saved csv file that contains log factorial for n <= 10000.}

\item{n}{A non-negative integer.}

\item{base}{The base of a logarithm.}
}
\description{
This is a relatively faster function to calculate the log factorial of a non-negative integer n. If n = 1, 
the result is 0; if n <= 10000, the result is read from a pre-saved log factorial sheet; if n > 10000, use
stirling's approximation.
}
