% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_gamma.R
\name{log_gamma}
\alias{log_gamma}
\title{A faster and accurate approximation to log factorial of a non-negative integer/real number}
\usage{
log_gamma(z)
}
\arguments{
\item{z}{A positive integer or real number.}
}
\description{
This function is more accuarte than Stirling's approximation. For more details, please check on
https://en.wikipedia.org/wiki/Lanczos_approximation. The log is natural base.
}
