% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prob_adaptive.R
\name{log_prob_adaptive}
\alias{log_prob_adaptive}
\title{An auxiliary function to mml_fixed_str()}
\usage{
log_prob_adaptive(data, targetIndex, condProbsAdpt)
}
\arguments{
\item{data}{A dataset whost variables are in numeric/integer format. Any categorical variables must be
converted into numeric/integer first.}

\item{targetIndex}{The target node's index in vars, whose Markov blanket we are interested in.}

\item{condProbsAdpt}{A conditional probabilities matrix that stores a variable's conditional probability
at each data point. This variable is either the target, or has more than one parent.}
}
\value{
The function outputs a single log probability value.
}
\description{
This function calculates log of the conditional probability for a variable given its parents. It is used
to obtain p(x|pa of x) in the numerator of the conditional prob p(T|Xs).
}
