% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb_false_finding.R
\name{mb_false_finding}
\alias{mb_false_finding}
\title{Evaluating Markov blanket learners according to false findings}
\usage{
mb_false_finding(mbt, mbl)
}
\arguments{
\item{mbt}{The true Markov blanket}

\item{mbl}{The learned Markov blanket}
}
\description{
This function returns the sum of false positives and negatives between the true and learned Markov blanket.
}
