% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_nb.R
\name{mle_nb}
\alias{mle_nb}
\title{MLE of NB parameters}
\usage{
mle_nb(data, vars, xIndices, yIndex, smoothing = 0.5)
}
\arguments{
\item{data}{A categorical data set.}

\item{vars}{A vector of all variables.}

\item{xIndices}{A vector of input variables indices.}

\item{yIndex}{The output variable index.}

\item{smoothing}{A value for parameter smoothing to avoid having 0 or 1 probabilities. 1 for laplacing 
smoothing, 0.5 for mml smoothing.}
}
\description{
Maximum likelihood of Naive Bayes parameters. Input parameters are always stored in front of output 
parameter in the returned list of parameters.
}
