% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_cpt_cach.R
\name{mml_cpt_cach}
\alias{mml_cpt_cach}
\title{An auxiliary function for mml_pt()}
\usage{
mml_cpt_cach(indexListPerNodePerValue, arities, sampleSize, vars, target,
  learnedMB)
}
\arguments{
\item{vars}{A vector of all variables.}

\item{target}{A taregt node.}

\item{learnedMB}{The learned mb of the target node}
}
\description{
Given a node x and its learned Markov blanket mb(x), this function pre-calculates the mml score of each node 
y in mb(x) given its potential parents. There are certain nodes that can't be parents of y, such as any 
subset that contains y, or any subset that doesn't contain the target node x, because if x is not a 
parent of y, then any parents of y must not exist in mb(x). This function helps speed up computing mmlcpt 
for an entire mbpt.
}
