% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_dag_fast.R
\name{mml_dag_fast}
\alias{mml_dag_fast}
\title{A function to calculate the mml score of a dag}
\usage{
mml_dag_fast(adjmtx, mmlCPTCach)
}
\arguments{
\item{adjmtx}{The adjacency matrix of a given dag.}

\item{mmlCPTCach}{The pre-calculated mml score of a node given its potential parents.}
}
\description{
This function calculates the mml score of a dag. The mml score of a dag is the sum of the mml score of a node given
its parents in the dag.
}
