% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_fixed_str_adaptive.R
\name{mml_fixed_str_adaptive}
\alias{mml_fixed_str_adaptive}
\title{MML for a fixed Markov blanket model using adaptive code approach}
\usage{
mml_fixed_str_adaptive(data, vars, arities, sampleSize, targetIndex,
  logProbTarget, cachedPXGivenT, probsMtx, str, mbIndices, cachedPXGivenY,
  cachInd)
}
\arguments{
\item{data}{A dataset whost variables are in numeric/integer format. Any categorical variables must be
converted into numeric/integer first.}

\item{vars}{A vector of all variables in data, in the same order as the column names of data.}

\item{arities}{A vector of variable arities in data, in the same order as the column names of data.}

\item{sampleSize}{The sample size. That is, the number of rows of data.}

\item{targetIndex}{The target node's index in vars, whose Markov blanket we are interested in.}

\item{logProbTarget}{Log of the probability of the target.}

\item{cachedPXGivenT}{cachedPXGivenT}

\item{probsMtx}{probsMtx}

\item{str}{A fixed Markov blanket model stores in the bnlearn dag format.}

\item{mbIndices}{The indices of potential Markov blanket variables.}

\item{cachedPXGivenY}{cachedPXGivenY}

\item{cachInd}{cachInd}
}
\value{
The function outputs the message length of a fixed structure.
}
\description{
This function calculates the mml score of the target under a fixed Markov blanket model. The structure
can be a general DAG. This returns the message length without stating the mml estimate of the parameters.
}
\keyword{This}
\keyword{cond_probs_adaptive().}
\keyword{dependencies}
\keyword{function}
\keyword{has}
\keyword{on}
