% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_nb_adaptive.R
\name{mml_nb_adaptive}
\alias{mml_nb_adaptive}
\title{MML Naive Bayes using adaptive code approach (faster version)}
\usage{
mml_nb_adaptive(data, arities, sampleSize, targetIndex, logProbTarget,
  cachedPXGivenT, chIndices)
}
\arguments{
\item{data}{A dataset whost variables are in numeric/integer format. Any categorical variables must be
converted into numeric/integer first.}

\item{arities}{A vector of variable arities.}

\item{sampleSize}{The sample size. That is, the number of rows of data.}

\item{targetIndex}{The target variable (or parent node) index.}

\item{logProbTarget}{Log of the probability of the target variable.}

\item{cachedPXGivenT}{Pre-calculated conditional probability of each node given its parents. In the case of Naive Bayes
models, the only parent is the target variable.}

\item{chIndices}{A vector of indices for the Xs (or child nodes).}
}
\description{
This function calculates the mml score of a NB model using the adaptive code approach. It is much simpler than using
MML87 that involves complex fisher calculation. The output is the mml score of a NB without adding the extra bit
for each parameter as what the multi-state mml adaptive code does, because we don't know the complete message length
from MML87 for NB. But maybe this is ok, because we are not interested in the mml estimation of parameters.
}
