% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_pt.R
\name{mml_pt}
\alias{mml_pt}
\title{A function to calculate the mml score of a polytree}
\usage{
mml_pt(mmlCPTCach, mbptsList, target, learnedMB)
}
\arguments{
\item{mmlCPTCach}{A matrix of pre-calculated mml score of each node in a mb given its possible parents. This is done
using the mml_cpt_cach() function.}

\item{mbptsList}{A list of all possible mbpts from mb size 0 up to 7.}

\item{target}{A taregt node.}

\item{learnedMB}{The learned mb of the target node.}
}
\description{
This function calculates the mml score of a polytree. It searches through the space of all possible polytrees using 
mml_cpt as the objective function.
}
