% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mml_with_parents_fast.R
\name{mml_with_parents_fast}
\alias{mml_with_parents_fast}
\title{Auxiliary function to mml_cpt_fast}
\usage{
mml_with_parents_fast(indexListPerNodePerValue, cachedIndicesList, arities,
  sampleSize, parentsIndices, targetIndex, alpha, statingPara)
}
\arguments{
\item{indexListPerNodePerValue}{A list of stored indices for each value of each node. It is obtained
by the function count_occurance().}

\item{cachedIndicesList}{A vector of indices stored to speed up calculations.}

\item{arities}{A vector of varaible arities.}

\item{sampleSize}{Sample size of a given data set.}

\item{parentsIndices}{Indices of parents nodes.}

\item{targetIndex}{Index of the target node.
#' @param alpha A vector of concentration parameters for a Dirichlet distribution. Range is from zeor to positive infinity,
length is equal to the arity of the target variable.}

\item{statingPara}{If TRUE MML estimate of the parameters are also stated with extra 0.5log(pi*e/6) per parameter,
otherwise 0.}
}
\description{
This is an auxiliary function to mml_cpt_fast. It calculates the mml score of a node given its parents.
}
