% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_len_add.R
\name{msg_len_add}
\alias{msg_len_add}
\title{Auxiliary function to msg_len_ave()}
\usage{
msg_len_add(x, y)
}
\arguments{
\item{x}{The message length of one model.}

\item{y}{The message length of another model.}
}
\value{
The function outputs the averaged message length in natural log.
}
\description{
This function uses message lengths to calculate the weighted averaged probability and returns an averaged 
message length based on this weighted averaged probability.
}
