% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_len_ave.R
\name{msg_len_ave}
\alias{msg_len_ave}
\title{Weighted averaged message length}
\usage{
msg_len_ave(l)
}
\arguments{
\item{l}{A vector of message lengths.}
}
\value{
The function a weighted averaged message length in natrual log.
}
\description{
This function takes in a vector of message lengths, and returnes an weighted averaged message length.
}
\keyword{This}
\keyword{dependencies}
\keyword{function}
\keyword{has}
\keyword{msg_len_add().}
\keyword{on}
