% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg_len_weighted_avg.R
\name{msg_len_weighted_avg}
\alias{msg_len_weighted_avg}
\title{Weighted average message length}
\usage{
msg_len_weighted_avg(l, w)
}
\arguments{
\item{l}{A vector of more than one message lengths.}

\item{w}{A vector of weights, one for each model (structure). It can be a vector of uniform 
probabilities that indicates an uniform averaging over all models.}
}
\value{
The function outputs a weighted average message length in natural log.
}
\description{
This function calculates the weighted average over a vector of message lengths. It can either be 
uniform or non-uniform. It uses message lengths instead of probabilities to avoid computer under
flow when dealing with extreme small probability multiplications.
}
