% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_pars_cal.R
\name{multi_pars_cal}
\alias{multi_pars_cal}
\title{Auxiliary function to mml_with_parents() and mml_with_parents_fast()}
\usage{
multi_pars_cal(indexListPerNodePerValue, commonParentsIndices, arityChild,
  targetIndex, alpha)
}
\arguments{
\item{indexListPerNodePerValue}{A list of stored indices for each value of each node. It is obtained
by the function count_occurance().}

\item{commonParentsIndices}{commonParentsIndices}

\item{arityChild}{arityChild}

\item{targetIndex}{Index of the target node.}

\item{alpha}{A vector of concentration parameters for a Dirichlet distribution. Range is from zeor to positive infinity,
length is equal to the arity of the target variable.}
}
\description{
This is an auxiliary function to both mml_with_parents() and mml_with_parents_fast() when the target
node has multiple parents.
}
