% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_logit.R
\name{nll_logit}
\alias{nll_logit}
\title{Calculate the negative log likelihood of an entire data set}
\usage{
nll_logit(dataNumeric, betaDotX, xIndices, yIndex)
}
\arguments{
\item{dataNumeric}{The numeric format of the given data set. Variable values start from 0. Obtained by 
using the function factor2numeric().}

\item{betaDotX}{Inner product of logit parameters beta and a data point X.}

\item{xIndices}{Indices of the input variables.}

\item{yIndex}{Input of the output/target variable.}
}
\description{
This function calculate the negative log likelihood of an entire data set.
}
