% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll_nb.R
\name{nll_nb}
\alias{nll_nb}
\title{NLL NB for a single data point}
\usage{
nll_nb(dataPoint, pars, xIndices, yIndex)
}
\arguments{
\item{dataPoint}{A single data point in the given data set.}

\item{pars}{A list of MLE of NB parameters.}

\item{xIndices}{A vector of input variables' indices.}

\item{yIndex}{The index of the output variable.}
}
\description{
This is an auxiliary function to calculate the negative log likelihood of the Naive Bayes model for a 
single data point given mle of parameters. That is, the function returns partial value of nll, without
calculating log(px). The complete value of nll is calculated in mml_nb().
}
