% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_value_to_index.R
\name{node_value_to_index}
\alias{node_value_to_index}
\title{An auxiliary function to mml_nb_adaptive()}
\usage{
node_value_to_index(arities, values)
}
\arguments{
\item{arities}{A vector of arities of the Xs and Y. It is in the order of c(xIndices, yIndex).}

\item{value}{A vector of variable values in the order of c(xIndices, yIndex). The values are in numeric format.}
}
\description{
This function converts the variable values to an index in the set of all combinations. For example, the values 
c("A","A","A") is converted into 1, c("B","A","A") is converted to 2.
}
