% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric2categorical.R
\name{numeric2categorical}
\alias{numeric2categorical}
\title{A function to convert numerical variables to categorical}
\usage{
numeric2categorical(data)
}
\arguments{
\item{data}{A numerical data set where each variable contains values such as 0, 1, 2, ...}
}
\description{
This function converts numerical variables to categorical (nominal or ordinal) and add the unique values
as the levels of each variable. This function is used when the input data set is in numerical format, such
as the data sets for the standard real world Bayesian networks ASIAN, ALARM, etc.
}
