% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_ijk.R
\name{p_ijk}
\alias{p_ijk}
\title{Auxiliary function to fim_nb()}
\usage{
p_ijk(dataPoint, pars, xIndices, xIndex, yValue)
}
\arguments{
\item{dataPoint}{A single data point in the given data set.}

\item{pars}{A list of MLE of NB parameters.}

\item{xIndices}{A vector of input variables' indices.}

\item{xIndex}{The idnex of a particular input variable in colnames(data).}

\item{yValue}{The value of the output variable. The current version of mml_nb only takes binar variables, 
so the possible y values are {0, 1}.}
}
\description{
This is an auxiliary function to fim_nb(). It calculates p(x_ij|y=k) for a single data point and store it
in a vector for future use. It is mainly for speed up the calculation of fim_nb().
}
