% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcmb2list.R
\name{pcmb2list}
\alias{pcmb2list}
\title{Convert pcmb's mb results to a list}
\usage{
pcmb2list(output, vars, method)
}
\arguments{
\item{output}{The screen output of pcmb/iamb.}

\item{vars}{A vector of variables.}

\item{method}{Either "pcmb" or "iamb".}

\item{n}{Total number of edges could be safely removed. n > 1.}
}
\description{
This function converts the MB output of pcmb and iamb (in c++) to a list. PCMB
outputs MB results on screen when calling sysmtem() in R.
}
