% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerset.R
\name{powerset}
\alias{powerset}
\title{A function to generate the power set of a set of variables.}
\usage{
powerset(x)
}
\arguments{
\item{x}{A set of variables.}
}
\description{
This function enumerates the power set of a set of variables except the set of all variables.
}
