% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs_adaptive.R
\name{probs_adaptive}
\alias{probs_adaptive}
\title{An auxiliary function to mml_fixed_str().}
\usage{
probs_adaptive(data, arities, sampleSize, probsMtx, curIndex)
}
\arguments{
\item{data}{A dataset whost variables are in numeric/integer format. Any categorical variables must be
converted into numeric/integer first.}

\item{arities}{A vector of variable arities in data, in the same order as the column names of data.}

\item{sampleSize}{The sample size. That is, the number of rows of data.}

\item{probsMtx}{probsMtx}

\item{curIndex}{The current variable's index. The current variable can either be the target or
a different variable.}
}
\value{
The function outputs a matrix of probabilities with dimension arity(curIndex) = arity(target)
by sample size.
}
\description{
This function calculates probability at each data point for a variable that has no parent. Noticing
we only consider no parents variable when it is the target.
}
