% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_pijk.R
\name{prod_pijk}
\alias{prod_pijk}
\title{Auxiliary function to fim_nb()}
\usage{
prod_pijk(dataPoint, pars, xIndices, yValue)
}
\arguments{
\item{dataPoint}{A single data point in the given data set.}

\item{pars}{A list of MLE of NB parameters.}

\item{xIndices}{A vector of input variables' indices.}

\item{yValue}{The value of the output variable. The current version of mml_nb only takes binar variables, 
so the possible y values are {0, 1}.}
}
\description{
This is an auxiliary function to fim_nb(). It calculates \prod_j p(x_ij|y=k) for a single data point 
and store it in a vector for future use. It is mainly for speed up the calculation of fim_nb(). Notice 
that this function is almost the same as p_ijk(), except that this function calculte the product over
all input variables x_j.
}
