% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_mb_model.R
\name{rand_mb_model}
\alias{rand_mb_model}
\title{A function to generate a random Markov blanket model (structure)}
\usage{
rand_mb_model(target, mb)
}
\arguments{
\item{target}{The target varaible.}

\item{mb}{The potential Markov blanket of the target.}
}
\value{
The output of this function is a Markov blanket model.
}
\description{
This function randomly generate a Markov blanket model for a given target and Markov blanket variables.
}
