% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_mbp.R
\name{rand_mbp}
\alias{rand_mbp}
\title{A function to generate a random Markov blanket polytree}
\usage{
rand_mbp(target, mb, format = c("matrix", "dag"))
}
\arguments{
\item{target}{The target varaible.}

\item{mb}{The potential Markov blanket of the target.}
}
\value{
The output of this function is a Markov blanket polytree
}
\description{
This function randomly generate a Markov blanket polytree for a given target and Markov blanket variables.
}
\keyword{matrix2dag()}
