% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_nbr_model.R
\name{rand_nbr_model}
\alias{rand_nbr_model}
\title{Random neighbourhood model}
\usage{
rand_nbr_model(target, nbr)
}
\arguments{
\item{target}{The target variable.}

\item{nbr}{nbr}
}
\description{
Random neighbourhood model
}
