% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randAdjmtx.R
\name{randAdjmtx}
\alias{randAdjmtx}
\title{A function to generate random adjacency matrix}
\usage{
randAdjmtx(nvars, maxNPa, adjList = NULL)
}
\arguments{
\item{nvars}{The desired number of varaibles.}

\item{maxNPa}{The desired maximum number of parents, with minimum 0 parent.}

\item{adjList}{A list of user preferred adjacencies. The default is NULL, meaning no preferred adjacenceis.}
}
\description{
This function randomly generate a DAG based on the specified structure details. It is the same as the 
randDag() function, except the generated DAG is stored in its adjacency matrix format. The adjacency
matrix can be transferred into the bnlearn format DAG using the matrix2dag() function.
}
