% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randDag.R
\name{randDag}
\alias{randDag}
\title{A function to generate a random DAG}
\usage{
randDag(nvars, maxNPas)
}
\arguments{
\item{nvars}{The desired number of varaibles.}

\item{maxNPa}{The desired maximum number of parents, with minimum 0 parent.}
}
\value{
The output of this function is a DAG in the bnlearn format.
}
\description{
This function randomly generate a DAG based on the specified structure details. Not all pairs of variables are
connected in the resulting DAG, especially for small number of maxNPas.
}
