% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_bubble_permutation.R
\name{random_bubble_permutation}
\alias{random_bubble_permutation}
\title{Random bubble permutation}
\usage{
random_bubble_permutation(a, k)
}
\arguments{
\item{a}{A random vector or array.}

\item{k}{an integer parameter indicating how many permutations are allowed.}
}
\description{
This function randomly permutes an adjacent pair of nodes in a given node order.
}
