% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_permutation.R
\name{random_permutation}
\alias{random_permutation}
\title{Random permutation of a vector}
\usage{
random_permutation(a)
}
\arguments{
\item{a}{A random vector or array.}
}
\description{
This function randomly permutes a vector or array of length n based on the
Algorithm235.
}
