% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_permutation_with_randomness.R
\name{random_permutation_with_randomness}
\alias{random_permutation_with_randomness}
\title{Random permutation of a vector}
\usage{
random_permutation_with_randomness(a, percentage)
}
\arguments{
\item{a}{A random vector or array.}

\item{percentage}{The percentage of hold fixed elements in the array. It takes
real value between 0 and 1, closed interval.}
}
\description{
This function randomly permutes a vector or array of length n based on the
Algorithm235. It also takes an input that specifies the percentage of hold
fixed elements.
}
