% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{A safer version of sample()}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{A vector consists of at least one element.}

\item{...}{The same parameters as sample().}
}
\description{
A safer version of sample()
}
