% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sll2list.R
\name{sll2list}
\alias{sll2list}
\title{Convert sll MB output to a list of MBs}
\usage{
sll2list(filePath, vars)
}
\arguments{
\item{filePath}{the full path of a sll MB output file.}

\item{vars}{A vector of variables. The order is consistent with the colnames
of the data set used by sll.}
}
\description{
This is a function to convert the learned MB from sll, stored in a text file
to a list of MBs for easy evaluation.
}
