% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_vars.R
\name{substitute_vars}
\alias{substitute_vars}
\title{An auxiliary function for mml_pt()}
\usage{
substitute_vars(ls, x, z)
}
\arguments{
\item{ls}{A list of pre-saved mbpts with variables v_1, ..., v_n.}

\item{x}{The target node.}

\item{z}{A vector of new variable names.}
}
\description{
This function replaces variable names in the pre-saved mbpts with the new variable names provided.
}
