% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry_correction.R
\name{symmetry_correction}
\alias{symmetry_correction}
\title{A function to enforce the symmetry correction after Markov blankets have been learned}
\usage{
symmetry_correction(vars, mbList, rule)
}
\arguments{
\item{vars}{A vector of all variables.}

\item{mbList}{A list of learned Markov blankets.}

\item{rule}{A string argument takes either "Intersection" or "Union", indicating the two deterministic rules mentioned
above.}
}
\description{
This function enforces the symmetry check for each learned mb. That is, if a node x is is mb(y), but y isn't
in mb(x), then either add y into mb(x) or delete x from mb(y), depending on which deterministic rule to
follow. The two famous deterministic rules are the Intersection and Union rules. For MB discovery using MML, we prefer
to use the Intersection rule due to the high precision and low recall of mml_cpt.
}
