% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry_correction_adj_mtx.R
\name{symmetry_correction_adj_mtx}
\alias{symmetry_correction_adj_mtx}
\title{A function to enforce the symmetry correction after Markov blankets have been learned}
\usage{
symmetry_correction_adj_mtx(mtx, rule)
}
\arguments{
\item{mtx}{Adjacency matrix that stores the learned MBs.}

\item{rule}{A string argument takes either "union" or "intersection",
indicating the two deterministic rules.}
}
\description{
This function does the same enforcement as symmetry_correction, but the MBs are
saved in adj mtx.
}
